+++ packages/ui/react-ui-editor/src/components/TextEditor/extensions/automerge/plugin.ts
@@ -19,8 +19,9 @@ import { type Heads, type Prop } from '@dxos/automerge/automerge';
 
 import { type IDocHandle } from './handle';
 import { PatchSemaphore } from './semaphore';
-import { CursorConverter } from '../../components/TextEditor/extensions/cursor-converter';
 import get from 'lodash.get';
+import { cursorConverter } from './cursor-converter';
+import { CursorConverter } from '../cursor-converter';
 
 export type Value = {
   lastHeads: Heads;
@@ -110,45 +111,15 @@ export const automergePlugin = (handle: IDocHandle, path: Prop[]): AutomergePlug
     },
   );
 
-  const cursorConverter: CursorConverter = {
-    toCursor: (pos: number) => {
-      const doc = handle.docSync();
-      if(!doc) return ''
-
-      const value = get(doc, path);
-      if(typeof value === 'string' && value.length <= pos) {
-        return 'end';
-      }
-
-      // NOTE: Slice is needed because getCursor mutates the array.
-      return automerge.getCursor(doc, path.slice(), pos);
-    },
-    fromCursor: (cursor: string) => {
-      if (cursor === '') {
-        return 0;
-      }
-
-      const doc = handle.docSync();
-      if(!doc) {
-        return 0;
-      }
-
-      if(cursor === 'end') {
-        const value = get(doc, path);
-        if(typeof value === 'string') {
-          return value.length;
-        } else {
-          return 0
-        }
-      }
-
-      // NOTE: Slice is needed because getCursor mutates the array.
-      return automerge.getCursorPosition(doc, path.slice(), cursor);
-    },
-  };
+  
 
   return {
-    extension: [stateField, semaphoreFacet.of(semaphore), viewPlugin, CursorConverter.of(cursorConverter)],
+    extension: [
+      stateField,
+      semaphoreFacet.of(semaphore),
+      viewPlugin,
+      CursorConverter.of(cursorConverter(handle, path))
+    ],
   };
 };
