+++ packages/ui/react-ui-editor/src/components/TextEditor/extensions/automerge/update-codemirror.ts
