+++ packages/ui/react-ui-editor/src/components/TextEditor/extensions/index.ts
@@ -18,3 +18,4 @@ export * from './table';
 export * from './tasklist';
 export * from './tooltip';
 export * from './typewriter';
+export * from './automerge';
